# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory Management System API",
        "description": "Hello my name is Tresten Pool aspiring backend developer. I created this api for my advanced software engineering class",
        "contact": {
            "email": "trestenpool@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://ec2-54-90-227-153.compute-1.amazonaws.com/api/"
        }
    ],
    "paths": {
        "/device": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "Index search for devices",
                "description": "The below query parameters are OPTIONAL. Feel free to mix and match",
                "operationId": "getdeviceList",
                "parameters": [
                    {
                        "name": "pagination",
                        "in": "query"
                    },
                    {
                        "name": "product_id",
                        "in": "query"
                    },
                    {
                        "name": "manufacturer_id",
                        "in": "query"
                    },
                    {
                        "name": "serial_number",
                        "in": "query"
                    },
                    {
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid Body"
                    }
                }
            },
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Store new device",
                "operationId": "newDevice",
                "requestBody": {
                    "description": "Pass in the device info",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "device": {
                                        "properties": {
                                            "product_id": {
                                                "type": "integer"
                                            },
                                            "manufacturer_id": {
                                                "type": "integer"
                                            },
                                            "serial_number": {
                                                "type": "string",
                                                "default": ""
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid Body"
                    },
                    "409": {
                        "description": "Conflict"
                    }
                }
            }
        },
        "/device/{device_id}": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "Get device by ID",
                "operationId": "getSpecificDevice",
                "parameters": [
                    {
                        "name": "device_id",
                        "in": "path",
                        "description": "Device ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid Request"
                    },
                    "404": {
                        "description": "Device not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Devices"
                ],
                "summary": "Update an existing device",
                "operationId": "updateDevice",
                "parameters": [
                    {
                        "name": "device_id",
                        "in": "path",
                        "description": "Device ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Pass in the device info",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "device": {
                                        "properties": {
                                            "product_id": {
                                                "type": "integer"
                                            },
                                            "manufacturer_id": {
                                                "type": "integer"
                                            },
                                            "serial_number": {
                                                "type": "string",
                                                "default": ""
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid Body"
                    },
                    "404": {
                        "description": "Device not found"
                    },
                    "409": {
                        "description": "Conflict"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Devices"
                ],
                "summary": "Delete a device",
                "operationId": "deleteDevice",
                "parameters": [
                    {
                        "name": "device_id",
                        "in": "path",
                        "description": "Device ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid Request"
                    },
                    "404": {
                        "description": "Device not found"
                    }
                }
            }
        },
        "/device/{device_id}/file": {
            "get": {
                "tags": [
                    "Device Files"
                ],
                "summary": "Show all the files for the device",
                "operationId": "showAllFiles",
                "parameters": [
                    {
                        "name": "device_id",
                        "in": "path",
                        "description": "Device ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid Request"
                    },
                    "404": {
                        "description": "Device not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Device Files"
                ],
                "summary": "Add a file to a device",
                "operationId": "AddFile",
                "parameters": [
                    {
                        "name": "device_id",
                        "in": "path",
                        "description": "Device ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "file",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid Request"
                    },
                    "404": {
                        "description": "Device not found"
                    }
                }
            }
        },
...
    }
}